ccu/webconsole/requirements.txt
django
django-method-override
pillow
six
pytz
humanize
yowsup
argparse
python-nmap
